+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Fixed unicode encoding error when publishing boards.
+
 ## v0.1.0 (2022-06-20)
 
 This release mostly implements the Spring '83 spec revision 1, dated 2022-06-16.